class定義を.hに移動,
